Call in the correct context.
Currently this means 'the context of the innermost enclosing sub', so
padblks will appear transparent to wantarray. I'm not sure yet what to
do about the return values: if we support supplying the return value for
arbitrary blocks, then wantarray really ought to reflect that block,
rather than the whole sub.