Allow keyword plugins to parse Perl syntax.
The way this works is that the keyword plugin returns
KEYWORD_PLUGIN_{BLOCK,COND}, which asks perly.y to parse either a block
or a ()-delimited expression for us. When it's finished with that, it
calls the keyword plugin again, so it can insert another PLUG* token if
it wants to. This second call is resposible for taking the results of
the parse and incorporating them into the returned optree. This makes
constructions like

    try {...}
    catch (...) {...}
    finally {...}

relatively easy to implement correctly.

Bugs: currently, the second call will necessarily start at the start of
the PL_keyword_plugin chain, which is not helpful. There is no easy way
for plugins to say 'this piece belongs to me' while handling recursion
properly. I think the best fix for this would be to replace the
PL_keyword_plugin hook with an AV of callbacks, so the parser can call
the correct entry directly the second time. (It's also somewhat cleaner
than relying on extensions to correctly call the next hook in the
chain.)