Replace PL_keyword_plugin with an AV.
This means making it an intrpvar instead of a true global, but I think
that's right anyway: installing a plugin hook in one interpreter
shouldn't affect another.

This means we get individual pointers to all the registered plugins, and
we can call back to the right one from perly.y.