Add error hooks, along the lines of the blockhooks.
This allows flexible hooking into eval/die/warn. The hooks can be used
as a cleaner alternative to $SIG{__WARN__} and __DIE__ (cleaner since
many hooks can be registered rather than just one), and also to keep
track of whether the current eval has failed yet or not. This can help
modules (such as DBIx::Class) that were relying on testing $@ in DESTROY
to determine this reliably.